+++ packages/ceramic-core/src/store/__tests__/state-store-integration.test.ts
@@ -8,6 +8,7 @@ import { TileDoctype, TileDoctypeHandler } from "@ceramicnetwork/ceramic-doctype
 import {PinStore} from "../pin-store";
 import {PinStoreFactory} from "../pin-store-factory";
 import { DID } from 'dids'
+import DocID from "@ceramicnetwork/docid"
 
 import { Resolver } from "did-resolver"
 import ThreeIdResolver from '@ceramicnetwork/3id-did-resolver'
@@ -200,7 +201,9 @@ describe('Level data store', () => {
   it('pins document correctly without IPFS pinning', async () => {
     const genesis = await TileDoctype.makeGenesis({ content: initialContent, metadata: { owners, tags: ['3id'] } }, context)
     const genesisCid = await dispatcher.storeRecord(genesis)
-    const doc = await Document.create(genesisCid, doctypeHandler, dispatcher, store, context)
+    const docId = new DocID('tile', genesisCid)
+    const doc = await Document.create(docId, doctypeHandler, dispatcher, store, context)
+
     await anchorUpdate(doc.doctype)
 
     let docState = await store.stateStore.load(doc.id)
@@ -216,7 +219,8 @@ describe('Level data store', () => {
   it('pins not anchored document correctly with IPFS pinning', async () => {
     const genesis = await TileDoctype.makeGenesis({ content: initialContent, metadata: { owners, tags: ['3id'] } }, context)
     const genesisCid = await dispatcher.storeRecord(genesis)
-    const doc = await Document.create(genesisCid, doctypeHandler, dispatcher, store, context, {
+    const docId = new DocID('tile', genesisCid)
+    const doc = await Document.create(docId, doctypeHandler, dispatcher, store, context, {
       applyOnly: true, skipWait: true,
     })
 
@@ -233,7 +237,8 @@ describe('Level data store', () => {
   it('pins document correctly with IPFS pinning', async () => {
     const genesis = await TileDoctype.makeGenesis({ content: initialContent, metadata: { owners, tags: ['3id'] } }, context)
     const genesisCid = await dispatcher.storeRecord(genesis)
-    const doc = await Document.create(genesisCid, doctypeHandler, dispatcher, store, context)
+    const docId = new DocID('tile', genesisCid)
+    const doc = await Document.create(docId, doctypeHandler, dispatcher, store, context)
     await anchorUpdate(doc.doctype)
 
     let docState = await store.stateStore.load(doc.id)
@@ -249,7 +254,8 @@ describe('Level data store', () => {
   it('removes pinned document', async () => {
     const genesis = await TileDoctype.makeGenesis({ content: initialContent, metadata: { owners, tags: ['3id'] } }, context)
     const genesisCid = await dispatcher.storeRecord(genesis)
-    const doc = await Document.create(genesisCid, doctypeHandler, dispatcher, store, context)
+    const docId = new DocID('tile', genesisCid)
+    const doc = await Document.create(docId, doctypeHandler, dispatcher, store, context)
     await anchorUpdate(doc.doctype)
 
     await store.add(doc.doctype)
@@ -262,7 +268,8 @@ describe('Level data store', () => {
   it('skips removing unpinned document', async () => {
     const genesis = await TileDoctype.makeGenesis({ content: initialContent, metadata: { owners, tags: ['3id'] } }, context)
     const genesisCid = await dispatcher.storeRecord(genesis)
-    const doc = await Document.create(genesisCid, doctypeHandler, dispatcher, store, context)
+    const docId = new DocID('tile', genesisCid)
+    const doc = await Document.create(docId, doctypeHandler, dispatcher, store, context)
     await anchorUpdate(doc.doctype)
 
     await store.rm(doc.id)
@@ -272,7 +279,8 @@ describe('Level data store', () => {
   it('lists pinned documents', async () => {
     const genesis = await TileDoctype.makeGenesis({ content: initialContent, metadata: { owners, tags: ['3id'] } }, context)
     const genesisCid = await dispatcher.storeRecord(genesis)
-    const doc = await Document.create(genesisCid, doctypeHandler, dispatcher, store, context)
+    const docId = new DocID('tile', genesisCid)
+    const doc = await Document.create(docId, doctypeHandler, dispatcher, store, context)
     await anchorUpdate(doc.doctype)
 
     await store.add(doc.doctype)
@@ -296,7 +304,8 @@ describe('Level data store', () => {
   it('lists empty for unpinned document', async () => {
     const genesis = await TileDoctype.makeGenesis({ content: initialContent, metadata: { owners, tags: ['3id'] } }, context)
     const genesisCid = await dispatcher.storeRecord(genesis)
-    const doc = await Document.create(genesisCid, doctypeHandler, dispatcher, store, context)
+    const docId = new DocID('tile', genesisCid)
+    const doc = await Document.create(docId, doctypeHandler, dispatcher, store, context)
     await anchorUpdate(doc.doctype)
 
     const pinned = []